+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/DeployModule.java
@@ -9,13 +9,11 @@ import lombok.NonNull;
 
 @Getter
 @EqualsAndHashCode(callSuper = true)
-public class DeployModule extends Payload {
+public class DeployModule extends PayloadUnknownCost {
     /**
      * A compiled Smart Contract Module in WASM with source and version.
      */
     private final WasmModule module;
-
-    private UInt64 maxEnergyCost;
 
 
     private DeployModule(@NonNull final WasmModule module) {
@@ -34,11 +32,6 @@ public class DeployModule extends Payload {
     }
 
     @Override
-    protected UInt64 getTransactionTypeCost() {
-        return this.maxEnergyCost;
-    }
-
-    @Override
     public TransactionType getTransactionType() {
         return TransactionType.DEPLOY_MODULE;
     }
@@ -46,10 +39,5 @@ public class DeployModule extends Payload {
     @Override
     protected byte[] getRawPayloadBytes() {
         return module.getBytes();
-    }
-
-    public DeployModule withMaxEnergyCost(UInt64 maxEnergyCost) {
-        this.maxEnergyCost = maxEnergyCost;
-        return this;
     }
 }